pyproject.toml
[project]
name = "class-elective-advisor-491"
version = "0.1.0"
description = ""
authors = [
    {name = "Matthew Dobley",email = "mkmshopping@sbcglobal.net"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langgraph (>=0.6.6,<0.7.0)",
    "langchain-openai (>=0.3.31,<0.4.0)",
    "langchain-core (>=0.3.74,<0.4.0)",
    "sphinx (>=8.2.3,<9.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-community (>=0.3.29,<0.4.0)",
    "tiktoken (>=0.11.0,<0.12.0)"
]

[tool.poetry]
packages = [{include = "class_elective_advisor_491", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
